Add files via upload
The text classification is one of the main tasks of computational linguistics because it combines
many other issues such as topic identification, author identification, sentiment analysis, and so on.
Content analysis of telecommunications networks is critical to ensuring information security and
public security. The text may contain illegal information (including data related to terrorism, drug
trafficking, organized protests, and large-scale riots). This article gives an overview of how to classify
text. The purpose of this study is to compare the latest methods of solving text classification
problems, identify trends, and select the best algorithms to use in research and commercial problems.
The latest and most well-known approach to text classification is based on machine learning
techniques. To choose a particular classification method, you need to take into account the properties
of each algorithm. This article describes the most common algorithms, the experiments
performed using them, and the results of these experiments. This survey was created between 2011
and 2016 and is based on scientific publications published on the Internet that are highly regarded
by the scientific community. This article analyzes and compares various classification methods
with characteristics such as precision, recall, execution time, possible algorithms in incremental
mode, range of background information required for classification, and language independence. It
contains.
I was hired by Nexocode as a junior data scientist, and my aim is to create end to end pipeline to
detect spam messages. I will try to perform all the necesary steps in data cleaning and analyzing
in order to achieve best outcome